String convertables categrories moved
