+++ Source/JSONJoy.swift
@@ -164,3 +164,24 @@ open class JSONDecoder {
 public protocol JSONJoy {
     init(_ decoder: JSONDecoder) throws
 }
+
+extension JSONDecoder: CustomStringConvertible {
+    public var description: String {
+        if let value = value {
+            return String(describing: value)
+        } else {
+            return String(describing: value)
+        }
+    }
+}
+
+extension JSONDecoder: CustomDebugStringConvertible {
+    public var debugDescription: String {
+        if let value = value {
+            return String(reflecting: value)
+        } else {
+            return String(reflecting: value)
+        }
+    }
+}
+
